requirements.txt
html2text
beautifulsoup4

python-dateutil==2.8.2
python-dotenv==1.0.1
regex==2023.12.25
dateparser==1.2.0
feedparser==6.0.11

loguru~=0.7.2
Markdown~=3.5.2

requests==2.31.0
urllib3==1.26.6

google-genai
openai==1.12.0




